added node merging for the purposes of partial param files changed the progress bar to real-time-based updates

+++ cg/include/cg/amino/sync_data.h
@@ -87,7 +87,7 @@ template <typename E> struct sync_data_auto_expr : public sync_data_expr<E> {
   AUTO_EXPR_BODY(back, side_all, side_polar, side_hydrophobic)
 };
 
-using sync_data_base = nitro::tuple_wrapper<int8_t, int8_t, int8_t, int8_t>;
+using sync_data_base = nitro::tuple_wrapper<int16_t, int16_t, int16_t, int16_t>;
 
 class sync_data : public sync_data_auto_expr<sync_data>, public sync_data_base {
 public:
